+++ lib/req/application.ex
@@ -6,7 +6,12 @@ defmodule Req.Application do
   @impl true
   def start(_type, _args) do
     children = [
-      {Finch, name: Req.Finch}
+      Supervisor.child_spec({Finch, name: Req.FinchHTTP1, pools: %{default: [protocol: :http1]}},
+        id: Req.FinchHTTP1
+      ),
+      Supervisor.child_spec({Finch, name: Req.FinchHTTP2, pools: %{default: [protocol: :http2]}},
+        id: Req.FinchHTTP2
+      )
     ]
 
     Supervisor.start_link(children, strategy: :one_for_one)